requirements.txt
fastapi
uvicorn
torch
diffusers
transformers
accelerate
safetensors
pillow